+++ allensdk/test/core/test_mouse_connectivity_notebook.py
@@ -38,19 +38,18 @@ import pytest
 import os
 from allensdk.test_utilities.temp_dir import fn_temp_dir
 
-@pytest.mark.skipif(True,
+@pytest.mark.skipif(os.getenv('TEST_COMPLETE') != 'true',
                     reason="partial testing")
 def test_notebook(fn_temp_dir):
-
 
     # coding: utf-8
 
     # ## Mouse Connectivity
-    #
+    # 
     # This notebook demonstrates how to access and manipulate data in the Allen Mouse Brain Connectivity Atlas. The `MouseConnectivityCache` AllenSDK class provides methods for downloading metadata about experiments, including their viral injection site and the mouse's transgenic line. You can request information either as a Pandas DataFrame or a simple list of dictionaries.
-    #
+    # 
     # An important feature of the `MouseConnectivityCache` is how it stores and retrieves data for you. By default, it will create (or read) a manifest file that keeps track of where various connectivity atlas data are stored. If you request something that has not already been downloaded, it will download it and store it in a well known location.
-    #
+    # 
     # Download this notebook in .ipynb format <a href='mouse_connectivity.ipynb'>here</a>.
 
     # In[1]:
@@ -61,7 +60,7 @@ def test_notebook(fn_temp_dir):
     # the data that has already been downloaded onto the hard drives.
     # If you supply a relative path, it is assumed to be relative to your
     # current working directory.
-    mcc = MouseConnectivityCache(manifest_file=os.path.join(fn_temp_dir, 'connectivity/mouse_connectivity_manifest.json'))
+    mcc = MouseConnectivityCache(manifest_file='connectivity/mouse_connectivity_manifest.json')
 
     # open up a list of all of the experiments
     all_experiments = mcc.get_experiments(dataframe=True)
@@ -87,11 +86,11 @@ def test_notebook(fn_temp_dir):
 
 
     # As a convenience, structures are grouped in to named collections called "structure sets". These sets can be used to quickly gather a useful subset of structures from the tree. The criteria used to define structure sets are eclectic; a structure set might list:
-    #
+    # 
     # * structures that were used in a particular project.
     # * structures that coarsely partition the brain.
     # * structures that bear functional similarity.
-    #
+    # 
     # or something else entirely. To view all of the available structure sets along with their descriptions, follow this [link](http://api.brain-map.org/api/v2/data/StructureSet/query.json). To see only structure sets relevant to the adult mouse brain, use the StructureTree:
 
     # In[3]:
@@ -136,7 +135,7 @@ def test_notebook(fn_temp_dir):
 
 
     # ## Structure Signal Unionization
-    #
+    # 
     # The ProjectionStructureUnionizes API data tells you how much signal there was in a given structure and experiment. It contains the density of projecting signal, volume of projecting signal, and other information. `MouseConnectivityCache` provides methods for querying and storing this data.
 
     # In[6]:
@@ -167,7 +166,7 @@ def test_notebook(fn_temp_dir):
 
     dense_unionizes = structure_unionizes[ structure_unionizes.projection_density > .5 ]
     large_unionizes = dense_unionizes[ dense_unionizes.volume > .5 ]
-    large_structures = pd.DataFrame(structure_tree.node(large_unionizes.structure_id))
+    large_structures = pd.DataFrame(structure_tree.nodes(large_unionizes.structure_id))
 
     print("%d large, dense, cortical, non-injection unionizes, %d structures" % ( len(large_unionizes), len(large_structures) ))
 
@@ -182,9 +181,10 @@ def test_notebook(fn_temp_dir):
     # In[9]:
 
     import numpy as np
+    import matplotlib.pyplot as plt
     import warnings
     warnings.filterwarnings('ignore')
-    
+
     visp_experiment_ids = [ e['id'] for e in visp_experiments ]
     ctx_children = structure_tree.child_ids( [isocortex['id']] )[0]
 
@@ -197,17 +197,36 @@ def test_notebook(fn_temp_dir):
     column_labels = [ c['label'] for c in pm['columns'] ] 
     matrix = pm['matrix']
 
+    fig, ax = plt.subplots(figsize=(15,15))
+    heatmap = ax.pcolor(matrix, cmap=plt.cm.afmhot)
 
+    # put the major ticks at the middle of each cell
+    ax.set_xticks(np.arange(matrix.shape[1])+0.5, minor=False)
+    ax.set_yticks(np.arange(matrix.shape[0])+0.5, minor=False)
+
+    ax.set_xlim([0, matrix.shape[1]])
+    ax.set_ylim([0, matrix.shape[0]])          
+
+    # want a more natural, table-like display
+    ax.invert_yaxis()
+    ax.xaxis.tick_top()
+
+    ax.set_xticklabels(column_labels, minor=False)
+    ax.set_yticklabels(row_labels, minor=False)
 
     # ## Manipulating Grid Data
-    #
-    # The `MouseConnectivityCache` class also helps you download and open every experiment's projection grid data volume. By default it will download 25um volumes, but uou could also download data at other resolutions if you prefer (10um, 50um, 100um).
-    #
-    # This demonstrates how you can load the projection density for a particular experiment. It also shows how to download the template volume to which all grid data is registered. Voxels in that template have been structurally annotated by neuroanatomists and stored in a separate annotation volume image.
+    # 
+    # The `MouseConnectivityCache` class also helps you download and open every experiment's projection grid data volume. By default it will download 25um volumes, but you could also download data at other resolutions if you prefer (10um, 50um, 100um).
+    # 
+    # This demonstrates how you can load the projection density for a particular experiment. It also shows how to download the template volume to which all grid data is registered. Voxels in that template have been structurally annotated by neuroanatomists and the resulting labels stored in a separate annotation volume image.
 
     # In[10]:
 
+    # we'll take this experiment - an injection into the primary somatosensory - as an example
     experiment_id = 181599674
+
+
+    # In[11]:
 
     # projection density: number of projecting pixels / voxel volume
     pd, pd_info = mcc.get_projection_density(experiment_id)
@@ -225,6 +244,64 @@ def test_notebook(fn_temp_dir):
     template, template_info = mcc.get_template_volume()
     annot, annot_info = mcc.get_annotation_volume()
 
+    # in addition to the annotation volume, you can get binary masks for individual structures
+    # in this case, we'll get one for the isocortex
+    cortex_mask, cm_info = mcc.get_structure_mask(315)
+
     print(pd_info)
     print(pd.shape, template.shape, annot.shape)
+
+
+    # Once you have these loaded, you can use matplotlib see what they look like.
+
+    # In[12]:
+
+    # compute the maximum intensity projection (along the anterior-posterior axis) of the projection data
+    pd_mip = pd.max(axis=0)
+    ind_mip = ind.max(axis=0)
+    inf_mip = inf.max(axis=0)
+
+    # show that slice of all volumes side-by-side
+    f, pr_axes = plt.subplots(1, 3, figsize=(15, 6))
+
+    pr_axes[0].imshow(pd_mip, cmap='hot', aspect='equal')
+    pr_axes[0].set_title("projection density MaxIP")
+
+    pr_axes[1].imshow(ind_mip, cmap='hot', aspect='equal')
+    pr_axes[1].set_title("injection density MaxIP")
+
+    pr_axes[2].imshow(inf_mip, cmap='hot', aspect='equal')
+    pr_axes[2].set_title("injection fraction MaxIP")
+
+
+    # In[13]:
+
+    # Look at a slice from the average template and annotation volumes
+
+    # pick a slice to show
+    slice_idx = 264
+
+    f, ccf_axes = plt.subplots(1, 3, figsize=(15, 6))
+
+    ccf_axes[0].imshow(template[slice_idx,:,:], cmap='gray', aspect='equal', vmin=template.min(), vmax=template.max())
+    ccf_axes[0].set_title("registration template")
+
+    ccf_axes[1].imshow(annot[slice_idx,:,:], cmap='gray', aspect='equal', vmin=0, vmax=2000)
+    ccf_axes[1].set_title("annotation volume")
+
+    ccf_axes[2].imshow(cortex_mask[slice_idx,:,:], cmap='gray', aspect='equal', vmin=0, vmax=1)
+    ccf_axes[2].set_title("isocortex mask")
+
+
+    # On occasion the TissueCyte microscope fails to acquire a tile. In this case the data from that tile should not be used for analysis. The data mask associated with each experiment can be used to determine which portions of the grid data came from correctly acquired tiles.
+    # 
+    # In this experiment, a missed tile can be seen in the data mask as a dark warped square. The values in the mask exist within [0, 1], describing the fraction of each voxel that was correctly acquired
+
+    # In[14]:
+
+    f, data_mask_axis = plt.subplots(figsize=(5, 6))
+
+    data_mask_axis.imshow(dm[81, :, :], cmap='hot', aspect='equal', vmin=0, vmax=1)
+    data_mask_axis.set_title('data mask')
+